Add constraints for nodeName and nodeSelector
This adds an admission controller that provides a boolean to allow or
prohibit the use of `nodeName` and `nodeSelector` fields in pod
requests. If `prohibitNodeTargetting` is set for the plugin in the
Openshift and Kubernetes master configuration, only accounts and
serviceaccounts which have the `pods/bind` permission and members of
groups that have the `pods/bind` permission will be permitted to create
Pods that target nodes, or to create resources that embed Pod templates
which target nodes.

prohibitNodeName/Selector now just one bool, prohibitNodeTargeting